Yarp: added an empty implementation of yarp motionControl class, an example of   initialization function.
The initialization should be moved somewhere else, becouse now it is self-destroing,
the yarp data structure should be reachable in other places of the code.

The code compiles and can be run to see the motionControl class get created and call
some function. An example of how to pass the configuration data is also present.

Try it and have fun!